refactor: Moved Channel localization to a common folder
