+++ services/channel/service/src/common/db/index.ts
@@ -2,4 +2,3 @@ export * from './apply-migrations';
 export * from './get-channel-field-for-plugin';
 export * from './get-playlist-field-for-plugin';
 export * from './migration-settings';
-export * from './set-is-localization-enabled-db-function';
