+++ services/channel/service/src/common/setup/index.ts
@@ -1,3 +1,1 @@
-export * from './register-image-types';
-export * from './register-localization-entity-definitions';
-export * from './register-video-cue-point-types';
+export * from './register-types';
