+++ services/channel/service/src/common/utils/index.ts
@@ -1,3 +1,2 @@
-export * from './is-localization-available';
 export * from './is-managed-service-enabled';
 export * from './token-utils';
