+++ services/channel/service/src/domains/index.ts
@@ -1,5 +1,3 @@
 export * from './channel';
-export * from './common';
 export * from './permission-definition';
 export * from './playlist';
-export * from './register-types';
