+++ services/channel/service/src/localization/index.ts
@@ -1,2 +1,4 @@
+export * from './is-localization-available';
 export * from './localizable-transactional-inbox-message-handler';
 export * from './localization-utils';
+export * from './set-is-localization-enabled-db-function';
